rgw: more header cleanup: remove code from rgw_rados.h
